requirements.txt
spacy==3.1.3
pdfplumber==0.5.28
PyMuPDF==1.18.19
pandas==1.3.3
openpyxl==3.0.9
PyPDF2==1.26.0
googletrans==4.0.0-rc1
tk==0.1.0
Flask==2.0.1
docx2pdf==0.1.7
python-docx==0.8.11
nltk==3.6.3
transformers==4.11.3
dash==2.0.0
streamlit==1.1.0
tqdm==4.62.3
concurrent.futures; python_version == "2.7"